move chain implementation into separate file
